hospitals/fhb/fhb_crawler_function/requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.41
botocore==1.34.41
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
jmespath==1.0.1
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
pytesseract==0.3.10
pytest==8.0.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
tomli==2.0.1
Unidecode==1.3.8
urllib3==1.26.18
